refactor(scalars): move sidebar to new UI folder
