+++ packages/design-system/src/ui/components/index.ts
@@ -9,3 +9,15 @@ export {
   DropdownShortcut,
   DropdownTrigger,
 } from "./dropdown/index.js";
+
+// sidebar
+export {
+  Sidebar,
+  SidebarProvider,
+  useSidebar,
+  type FlattenedNode,
+  type NodeStatus,
+  type SidebarIcon,
+  type SidebarNode,
+  type SidebarProps,
+} from "./sidebar/index.js";
