+++ packages/design-system/src/ui/components/sidebar/index.ts
@@ -1,2 +1,3 @@
 export * from "./sidebar.js";
+export * from "./subcomponents/sidebar-provider/index.js";
 export type * from "./types.js";
