+++ packages/design-system/src/ui/components/sidebar/mocked_tree.json
