+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-content-area.tsx
