+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-pinning-area.tsx
