+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-provider/sidebar-reducer.tsx
