+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-search/index.ts
