+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-search/search-result-counter.tsx
