+++ packages/design-system/src/ui/components/sidebar/subcomponents/sidebar-search/sidebar-search.tsx
@@ -4,8 +4,8 @@ import { useCallback, useRef } from "react";
 import CrossCircle from "../../../../../powerhouse/components/icon-components/CrossCircle.js";
 import Search from "../../../../../powerhouse/components/icon-components/Search.js";
 import Tabler from "../../../../../powerhouse/components/icon-components/Tabler.js";
-import { cn } from "../../../../lib/utils.js";
-import { Input } from "../../../fragments/index.js";
+import { Input } from "../../../../../scalars/components/fragments/index.js";
+import { cn } from "../../../../../scalars/lib/utils.js";
 import { useSidebar } from "../sidebar-provider/index.js";
 import { SearchNavigationArrow } from "./search-navigation-arrow.js";
 import { SearchResultCounter } from "./search-result-counter.js";
