+++ packages/design-system/src/ui/components/sidebar/subcomponents/status-icon.tsx
